+++ src/Api/Solver.test.tsx
@@ -1,5 +1,5 @@
 
-import { render, screen } from "@testing-library/react";
+import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import App from "../App";
 import { forComponentsToStopUpdating, forComponentsToUpdate } from "../utils";
@@ -34,45 +34,34 @@ export function currentStrategyIndex() {
    return Array.prototype.indexOf.call(parentChildren, currentStrategy)
 }
 
+test("No visuallyCurrentStrategy by default", () => {
+   expect(visuallyCurrentStrategy()).toBeUndefined()
+})
+
 test.skip("Stays at first strategy when board is invalid", async () => {
    await importBoard(BOARDS["Invalid board"])
-   userEvent.click(screen.getByRole("button", { name: "go" }))
-   await forComponentsToStopUpdating()
-   expect(currentStrategyIndex()).toBe(0)
+   userEvent.click(screen.getByRole("button", { name: "step" }))
+   userEvent.click(screen.getByRole("button", { name: "step" }))
+   await waitFor(() => expect(currentStrategyIndex()).toBe(0))
 })
 
-test.skip("Strategy index can increase", async () => {
+test.skip("Strategy index starts at 0", async () => {
    await importBoard(BOARDS["Simple sudoku"])
    userEvent.click(screen.getByRole("button", { name: "step" }))
-   await forComponentsToStopUpdating()
-   expect(currentStrategyIndex()).toBe(1)
+   await waitFor(() => expect(currentStrategyIndex()).toBe(0))
 })
 
-test.skip("No matter how fast the button is clicked, strategies will be done", async () => {
-   await importBoard(BOARDS["Very disconnected digits"])
+test.skip("Strategy index goes to 1 next", async () => {
+   await importBoard(BOARDS["Simple sudoku"])
    userEvent.click(screen.getByRole("button", { name: "step" }))
    userEvent.click(screen.getByRole("button", { name: "step" }))
-   await forComponentsToStopUpdating()
-   screen.debug(screen.getByRole("table"))
-   expect(currentStrategyIndex()).toBe(2)
+   await waitFor(() => expect(currentStrategyIndex()).toBe(1))
 })
 
 test.skip("After a strategy success, the index is 0 again", async () => {
    await importBoard(BOARDS["Simple sudoku"])
-
-   // Success!
    userEvent.click(screen.getByRole("button", { name: "step" }))
-   await forComponentsToStopUpdating()
-   screen.debug((visuallyCurrentStrategy() as HTMLElement).parentElement as HTMLElement)
-   expect(currentStrategyIndex()).toBe(1)
-
    userEvent.click(screen.getByRole("button", { name: "step" }))
-   await forComponentsToStopUpdating()
-   screen.debug((visuallyCurrentStrategy() as HTMLElement).parentElement as HTMLElement)
-   expect(currentStrategyIndex()).toBe(2)
-
-   // Goes back to 0
    userEvent.click(screen.getByRole("button", { name: "step" }))
-   await forComponentsToStopUpdating()
-   expect(currentStrategyIndex()).toBe(0)
+   await waitFor(() => expect(currentStrategyIndex()).toBe(0))
 })