+++ src/Api/Sudoku.ts
@@ -3,7 +3,7 @@ import { IndexToNine, SudokuDigits, TwoDimensionalArray } from "../Types"
 import PureSudoku from "./PureSudoku"
 
 export default class Sudoku extends PureSudoku {
-   cells: TwoDimensionalArray<Cell>
+   cells: TwoDimensionalArray<Cell | undefined | null>
    constructor (...args: ConstructorParameters<typeof PureSudoku>) {
       super(...args)
       if (this.data === undefined) {
@@ -25,7 +25,7 @@ export default class Sudoku extends PureSudoku {
             // super calls set() before this.cells is set
             if (this.cells !== undefined) {
                await new Promise(resolve => {
-                  this.cells[x][y].setCandidatesTo(candidates, () => resolve(undefined))
+                  this.cells[x][y]?.setCandidatesTo(candidates, () => resolve(undefined))
                })
             }
          }
@@ -36,13 +36,18 @@ export default class Sudoku extends PureSudoku {
     * This is currently used for initialization, but could also be used for updating
     * That's pretty complicated
     */
-   updateFromCell(cell: Cell) {
+   addCell(cell: Cell) {
       this.cells[cell.props.row][cell.props.column] = cell
       this.data[cell.props.row][cell.props.column] = cell.state.candidates
    }
 
+   removeCell(cell: Cell) {
+      this.cells[cell.props.row][cell.props.column] = undefined
+      // this.data[cell.props.row][cell.props.column] = cell.state.candidates
+   }
+
    clearCell(x: IndexToNine, y: IndexToNine) {
       this.data[x][y] = [1, 2, 3, 4, 5, 6, 7, 8, 9]
-      this.cells[x][y].clearCandidates()
+      this.cells[x][y]?.clearCandidates()
    }
 }